Fix fread() blocking until timeout in Stomp::readFrame()
The fread() loop in readFrame uses the default socket timeout
properties to determine how long to attempt to read 1024 bytes
from the socket. However if the frame size is > 1024
bytes then the fread will block on getting the final frame fragment.

Since it is desirable to have a high timeout for the stream_select
call, but not for the fread, I've set the stream timeout to 20,000
microseconds (0.2 seconds). The stream_select will still use the
explicit Stomp::_read_timeout_* properties, but the fread() will
never block for longer than 0.2 seconds.